% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipg_grade.R
\docType{data}
\name{ac_cols}
\alias{ac_cols}
\title{FSOT Domain Columns}
\format{
An object of class \code{character} of length 2.
}
\usage{
ac_cols
}
\description{
Column groups used in FSOT scoring for Aligned Content (AC), Assessment & Differentiation (AD), Student Practice (SP), and Teacher-Directed Instruction (TD).
}
\keyword{datasets}
