% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipg_grade.R
\docType{data}
\name{ca1_cols}
\alias{ca1_cols}
\title{Core Action Column Groups}
\format{
An object of class \code{character} of length 12.
}
\usage{
ca1_cols
}
\description{
Convenience vectors grouping IPG rubric items by Core Action (CA1, CA2, CA3) for scoring.
These include both character and numeric columns.
}
\keyword{datasets}
