% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\name{calc_nps}
\alias{calc_nps}
\title{Calculate nps score}
\usage{
calc_nps(x)
}
\arguments{
\item{x}{A vector of nps scores}
}
\value{
Returns the nps score
With formula `%` promoters - `%` detractors where promoters are 9 or 10 ratings and detractors are 0 to 6
}
\description{
Calculate nps score
}
