% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\name{check_email_approved}
\alias{check_email_approved}
\title{Verify email from vector}
\usage{
check_email_approved(email, approved_emails_list)
}
\arguments{
\item{email}{the email provided}

\item{approved_emails_list}{the email to check the list for}
}
\value{
TRUE or FALSE
}
\description{
Verifies that an email is in a list
}
