% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\name{check_email_domain}
\alias{check_email_domain}
\title{Verify email domain}
\usage{
check_email_domain(email, domain)
}
\arguments{
\item{email}{the email provided}

\item{domain}{the domain to check that it is from}
}
\value{
TRUE or FALSE
}
\description{
Verifies that an email is of a specified domain
}
