% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\name{contact_lead_graph}
\alias{contact_lead_graph}
\title{Contact Lead Graph Summary}
\usage{
contact_lead_graph(data)
}
\arguments{
\item{data}{the data to use}
}
\value{
prints a ggplot object
}
\description{
Returns a barchart for selections in the relevant questions of the contact lead survey
}
