% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\name{date_filter}
\alias{date_filter}
\title{Date filter}
\usage{
date_filter(data, date1, date2, dat_filter)
}
\arguments{
\item{data}{the dataframe to apply filter to}

\item{date1}{the min date}

\item{date2}{the max date}

\item{dat_filter}{Data column object to filter}
}
\value{
filtered data.frame
}
\description{
Efficiently filters data frame by two dates (inclusive),
for use in shiny apps
}
