% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\name{drop1}
\alias{drop1}
\title{HTML/CSS Button Content Expander}
\usage{
drop1(before = T, options, envir, name)
}
\arguments{
\item{before}{button default, collapsed/not collapsed}

\item{options}{unclear}

\item{envir}{also unclear}

\item{name}{chunk name}
}
\value{
html wrapper
}
\description{
Creates a button that will expand or hide content
}
