% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\name{file.path2}
\alias{file.path2}
\title{File path}
\usage{
file.path2(..., fsep = .Platform$file.sep)
}
\arguments{
\item{...}{The file path}

\item{fsep}{the file separation}
}
\value{
fp a file path
}
\description{
Gives the file path without double slash bug
}
