% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_makers.R
\name{find_highlight}
\alias{find_highlight}
\title{Word highlighting}
\usage{
find_highlight(string, n = 3, print = F)
}
\arguments{
\item{string}{the string to evaluate}

\item{n}{the number of words to find}

\item{print}{whether or not to print the highlighted words}
}
\value{
a vector of strings
}
\description{
Finds most common words in string
}
