% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\name{get_percent_positive_student_survey}
\alias{get_percent_positive_student_survey}
\title{Get % positive on any of the student surveys}
\usage{
get_percent_positive_student_survey(data)
}
\arguments{
\item{data}{the data to use}
}
\value{
a tibble
}
\description{
Uses column names and finds the % in each attributed to positive answers (or negative if reverse coded), summarises by prepost
}
