% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\name{grade_ipg}
\alias{grade_ipg}
\title{Grade IPG Data}
\usage{
grade_ipg(x, type = "character")
}
\arguments{
\item{x}{the data}

\item{type}{character or numeric}
}
\value{
a percentage of correct either by checking 3 or 4 or yes
}
\description{
function for grading different parts of the ipg forms
}
