% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\name{gt_percent_n}
\alias{gt_percent_n}
\title{GT or ggplot maker}
\usage{
gt_percent_n(
  df,
  column,
  custom_title,
  no_title = T,
  base_font = 10,
  heading_font = 14,
  custom_column_name = "",
  viz_type = "gt"
)
}
\arguments{
\item{df}{the data frame}

\item{column}{the column to get count and percent from}

\item{custom_title}{the title for the table}

\item{no_title}{make the table have no title}

\item{base_font}{overall table font size}

\item{heading_font}{title font size}

\item{custom_column_name}{a custom name for the column}

\item{viz_type}{gt by default, also has ggplot options like pie chart, waffle, or treemap}
}
\value{
a gt table
}
\description{
makes a gt table with percent and n colored
}
