% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl_themes.R
\name{gt_theme_tl}
\alias{gt_theme_tl}
\title{Create Teaching Lab theme to a gt table}
\usage{
gt_theme_tl(
  data,
  all_caps = F,
  align = "center",
  base_font = 16,
  heading_font = 20,
  ...
)
}
\arguments{
\item{data}{An existing gt table object}

\item{all_caps}{Whether or not to capitalize titles}

\item{align}{Align options are "left", "center", "right"}

\item{base_font}{the font size}

\item{heading_font}{the title font size}

\item{...}{Optional additional arguments to gt::table_options()}
}
\value{
Creates a gt theme as a pipeable function
}
\description{
Create Teaching Lab theme to a gt table
}
\examples{
mtcars \%>\% utils::head() \%>\% gt::gt() \%>\% tlShiny::gt_theme_tl()
}
