% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_makers.R
\name{gtable_extract_grob}
\alias{gtable_extract_grob}
\title{gtable_extract_grob}
\usage{
gtable_extract_grob(g, pattern = "guide-box")
}
\arguments{
\item{g, }{the gtable to extract the grob}

\item{pattern, }{grob name or pattern to match}
}
\value{
g, a grob matching the specified pattern
}
\description{
Helper function to extract a grob from gtables by name.
}
