% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_makers.R
\name{gtable_remove_grob}
\alias{gtable_remove_grob}
\title{gtable_remove_grob}
\usage{
gtable_remove_grob(g, pattern = "guide-box")
}
\arguments{
\item{g, }{gtable with the grob removed}

\item{pattern}{grob name or pattern to match}
}
\value{
g, with pattern removed.
}
\description{
Helper function to remove grobs by name, from gtables
}
