% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\name{html_wrap}
\alias{html_wrap}
\title{HTML Text Wrapping}
\usage{
html_wrap(string, n = 40)
}
\arguments{
\item{string}{the string}

\item{n}{the width of the string before a <br> tag}
}
\value{
the same string with <br> inserted at the requested interval
}
\description{
Takes a string and inserts <br> at the requested intervals
}
\examples{
html_wrap("a random string that has about 40 characters in it")
}
