% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipg_grade.R
\docType{data}
\name{ipg_character_cols}
\alias{ipg_character_cols}
\title{IPG Character Columns}
\format{
An object of class \code{character} of length 6.
}
\usage{
ipg_character_cols
}
\description{
Character column names from the Instructional Practice Guide (IPG) rubric.
These are typically binary values like "Yes"/"No" indicating presence of a practice.
}
\keyword{datasets}
