% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipg_grade.R
\docType{data}
\name{ipg_numeric_cols}
\alias{ipg_numeric_cols}
\title{IPG Numeric Columns}
\format{
An object of class \code{character} of length 51.
}
\usage{
ipg_numeric_cols
}
\description{
Numeric column names from the IPG rubric, scored from 1 to 4.
Higher scores indicate stronger alignment with the rubric.
}
\keyword{datasets}
