% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipg_grade.R
\docType{data}
\name{ipg_numeric_low_cols}
\alias{ipg_numeric_low_cols}
\title{IPG Numeric Low Columns}
\format{
An object of class \code{character} of length 2.
}
\usage{
ipg_numeric_low_cols
}
\description{
Numeric columns from the IPG rubric with a lower scale where 2 and 3 are considered "high."
}
\keyword{datasets}
