% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_makers.R
\name{know_assess_summary}
\alias{know_assess_summary}
\title{Knowledge Assessment Graph Summary}
\usage{
know_assess_summary(data, know_assess_filter)
}
\arguments{
\item{data}{the data}

\item{know_assess_filter}{the knowledge assessment to make plot for}
}
\value{
a ggplot
}
\description{
Creates a graph specifically for Knowledge Assessments in mid year reports
}
