% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_makers.R
\name{know_assess_summary_detailed}
\alias{know_assess_summary_detailed}
\title{Knowledge Assessment Graph Question-level Summary}
\usage{
know_assess_summary_detailed(data, know_assess)
}
\arguments{
\item{data}{the data}

\item{know_assess}{the knowledge assessment to make plot for}
}
\value{
a ggplot
}
\description{
Creates a graph specifically for Knowledge Assessments Scored by Question from Qualtrics
}
