% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_makers.R
\name{make_ipg_ela_summary_chart}
\alias{make_ipg_ela_summary_chart}
\title{Knowledge Assessment Graph Question-level Summary}
\usage{
make_ipg_ela_summary_chart(
  data,
  round = "Baseline (first observation of the year)"
)
}
\arguments{
\item{data}{the data}

\item{round}{the round of the IPG for which data should be pulled (can be one of "Baseline (first observation of the year)", "Mid-year (middle of service, if applicable)", "End of year (last observation of the year)", "Other", or "Ongoing")}
}
\value{
a ggplot
}
\description{
Creates a graph specifically for Knowledge Assessments Scored by Question from Qualtrics
}
