% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_makers.R
\name{make_ipg_math_summary_chart}
\alias{make_ipg_math_summary_chart}
\title{Math IPG Chart Summary Maker}
\usage{
make_ipg_math_summary_chart(round = "Baseline (first observation of the year)")
}
\arguments{
\item{round}{the round of the IPG for which data should be pulled (can be one of "Baseline (first observation of the year)", "Mid-year (middle of service, if applicable)", "End of year (last observation of the year)", "Other", or "Ongoing")}

\item{data}{the data}
}
\value{
a ggplot
}
\description{
Creates a chart to summarise the Math IPG based on the round selected
}
