% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_makers.R
\name{make_ipg_rubric_summary}
\alias{make_ipg_rubric_summary}
\title{IPG Rubric Summary}
\usage{
make_ipg_rubric_summary(data, rubric)
}
\arguments{
\item{data}{the data}

\item{rubric}{the ipg rubric to focus on}
}
\value{
a ggplot
}
\description{
Creates a chart to summarise the IPG Results for a specified rubric
}
