% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\name{make_overall_mindsets}
\alias{make_overall_mindsets}
\title{Teacher Mindsets Table}
\usage{
make_overall_mindsets(data)
}
\arguments{
\item{data}{the data}
}
\value{
a gt
}
\description{
Creates a table to summarise the results of the mindsets section of the educator survey
}
