% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\name{make_student_work_chart_circle}
\alias{make_student_work_chart_circle}
\title{Student Work Circle Charts}
\usage{
make_student_work_chart_circle(data)
}
\arguments{
\item{data}{the data for the plotter to use, should include all columns of interest}
}
\value{
a ggplot
}
\description{
Chart to show the % of people for student work that are on grade level
}
