% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\name{make_student_work_chart_people}
\alias{make_student_work_chart_people}
\title{Student Work People Chart}
\usage{
make_student_work_chart_people(data)
}
\arguments{
\item{data}{the data for the plotter to use, should include all columns of interest}
}
\value{
a ggplot
}
\description{
Chart to show the % of people in each scoring category for student work
}
