% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_makers.R
\name{make_teacher_curriculum_perceptions}
\alias{make_teacher_curriculum_perceptions}
\title{Teacher Curriculum Perceptions Graph}
\usage{
make_teacher_curriculum_perceptions(data)
}
\arguments{
\item{data}{the data}
}
\value{
a ggplot
}
\description{
Creates a chart to summarise the teacher perceptions of curriculum
}
