% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_makers.R
\name{make_teacher_curriculum_usage}
\alias{make_teacher_curriculum_usage}
\title{Teacher Curriculum Usage Graph}
\usage{
make_teacher_curriculum_usage(data)
}
\arguments{
\item{data}{the data}
}
\value{
a ggplot
}
\description{
Creates a chart to summarise the Teacher Curriculum Usage from the Educator Survey
}
