% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_makers.R
\name{make_teacher_lesson_usage}
\alias{make_teacher_lesson_usage}
\title{Teacher Use of Lessons Graph}
\usage{
make_teacher_lesson_usage(data)
}
\arguments{
\item{data}{the data}
}
\value{
a ggplot
}
\description{
Creates a chart to summarise the teacher use of lessons from district or school-adopted materials
}
