% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_makers.R
\name{make_teacher_perceptions_peer_relationships}
\alias{make_teacher_perceptions_peer_relationships}
\title{Teacher Perceptions Peer Relationships Graph}
\usage{
make_teacher_perceptions_peer_relationships(data)
}
\arguments{
\item{data}{the data}
}
\value{
a ggplot
}
\description{
Creates a chart to summarise the teacher perceptions of of peer relationships
}
