% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\name{make_teacher_perceptions_school_leaders}
\alias{make_teacher_perceptions_school_leaders}
\title{Teacher School Leader Perceptions Graph}
\usage{
make_teacher_perceptions_school_leaders(data)
}
\arguments{
\item{data}{the data}
}
\value{
a ggplot
}
\description{
Creates a chart to summarise the teacher perceptions of school leaders
}
