% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_df.R
\docType{data}
\name{na_df}
\alias{na_df}
\title{A list of NA type words to ignore in qualitative responses}
\format{
A character vector
}
\usage{
na_df
}
\description{
A list of NA type words to ignore in qualitative responses
}
\keyword{datasets}
