% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\name{neg_cond_filter}
\alias{neg_cond_filter}
\title{Negative Conditional Filter}
\usage{
neg_cond_filter(data, if_not_this, filter_this, dat_filter)
}
\arguments{
\item{data}{the dataframe to apply filter}

\item{if_not_this}{If value is not this}

\item{filter_this}{Filter for this}

\item{dat_filter}{Data column object to filter}
}
\value{
filtered data.frame
}
\description{
Conditionally filters value given that it is not the first parameter,
for use in shiny apps
}
