% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\docType{data}
\name{no_data_plot_currently}
\alias{no_data_plot_currently}
\title{No Data Plot}
\format{
An object of class \code{gg} (inherits from \code{ggplot}) of length 9.
}
\usage{
no_data_plot_currently
}
\description{
A plot that says no data available yet this year
}
\keyword{datasets}
