% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\name{no_data_plot_custom}
\alias{no_data_plot_custom}
\title{No Data Plot Adjustable Title}
\usage{
no_data_plot_custom(title = "No data available!")
}
\description{
A plot that says no data available by default, but can be adjusted to say whatever you want
}
