% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\name{ongoing_coaching_feedback_graph}
\alias{ongoing_coaching_feedback_graph}
\title{Ongoing Coaching feedback graph dependent on race and content area}
\usage{
ongoing_coaching_feedback_graph(data)
}
\arguments{
\item{data}{the data to use}
}
\value{
prints a ggplot object
}
\description{
Returns a barchart for selections in the relevant questions of the end of coaching survey
}
