% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipg_grade.R
\docType{data}
\name{prek_cols}
\alias{prek_cols}
\title{PreK Rubric Columns}
\format{
An object of class \code{character} of length 19.
}
\usage{
prek_cols
}
\description{
Convenience vector combining all PreK rubric items (Language + Literacy).
}
\keyword{datasets}
