% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipg_grade.R
\docType{data}
\name{prek_lang_cols}
\alias{prek_lang_cols}
\title{PreK Language Columns}
\format{
An object of class \code{character} of length 10.
}
\usage{
prek_lang_cols
}
\description{
Character column names from the PreK rubric for Language.
Values are Likert-type, with "Always" and "Often" treated as proficient.
}
\keyword{datasets}
