% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipg_grade.R
\docType{data}
\name{prek_lit_cols}
\alias{prek_lit_cols}
\title{PreK Literacy Columns}
\format{
An object of class \code{character} of length 9.
}
\usage{
prek_lit_cols
}
\description{
Character column names from the PreK rubric for Literacy.
Values are Likert-type, with "Always" and "Often" treated as proficient.
}
\keyword{datasets}
