% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_makers.R
\name{quote_viz}
\alias{quote_viz}
\title{Quote Visualization}
\usage{
quote_viz(
  data,
  text_col = colnames(data)[1],
  viz_type = "gt",
  custom_highlight = NULL,
  n = 3,
  print = T,
  width = 60,
  title = NULL,
  suppress_warnings = T,
  align = "center",
  save = T,
  ...
)
}
\arguments{
\item{data}{the dataframe}

\item{text_col}{columns to create table for}

\item{viz_type}{ggplot or gt visualization}

\item{custom_highlight}{a vector, optional custom highlighting}

\item{n}{integer, number of words to auto-highlight}

\item{print}{T, whether or not to print the highlighted words to console}

\item{width}{The width of the table generated}

\item{title}{the title of the ggplot or gt}

\item{suppress_warnings}{T/F suppression of warnings}

\item{align}{the table alignment: "left", "center", "right"}

\item{...}{Arguments passed onto the gt table}
}
\value{
a ggplot/gt that visualizes text
}
\description{
takes a dataframe and makes a gt table or ggplot that shows a quote
}
\examples{
\dontrun{
df <- TeachingLab::survey_monkey
colnames(df)[1] <- "What learning are you excited to try?"
quote_viz(
  data = df,
  text_col = "What learning are you excited to try?",
  viz_type = "gt",
  title = "Responses from Survey Monkey"
)
}
}
