% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\name{remove_empty_tl}
\alias{remove_empty_tl}
\title{Remove Empty Rows}
\usage{
remove_empty_tl(data)
}
\arguments{
\item{data}{the dataframe to apply filter}
}
\value{
a dataframe with no empty rows
}
\description{
Filters out all rows that contain entirely NA
}
