% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\name{session_feedback_graph}
\alias{session_feedback_graph}
\title{End of session feedback graph dependent on race and content area}
\usage{
session_feedback_graph(data)
}
\arguments{
\item{data}{the data to use}
}
\value{
prints a ggplot object
}
\description{
Returns a barchart for selections in the relevant questions of the end of coaching survey
}
