% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Calculate the square of a number}
\usage{
square(x)
}
\arguments{
\item{x}{A numeric input.}
}
\value{
The square of the input.
}
\description{
This function computes the square of a numeric input.
}
\examples{
tlShiny::square(2)
tlShiny::square(3.5)
}
