% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\name{student_bar_chart}
\alias{student_bar_chart}
\title{Student percent agree/strongly agree}
\usage{
student_bar_chart(
  data,
  col_select,
  agree_select,
  string_remove,
  title,
  legend_position = c(0.8, 0.25)
)
}
\arguments{
\item{data}{the data for the plotter to use, should include all columns of interest}

\item{col_select}{the columns to select with `tidyselect::contains`}

\item{agree_select}{the type of agree/strongly agree to select, for example also often/always}

\item{string_remove}{NULL by default, provides an optional string removal}

\item{title}{the title for the plot}

\item{legend_position}{c(0.8, 0.25) by default, adjustable}
}
\value{
a ggplot object
}
\description{
Automatically dodged bar chart for student data
}
