% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsites.R
\name{subsites}
\alias{subsites}
\title{The current subsites selection}
\format{
A vector of subsite names
}
\source{
\url{https://docs.google.com/spreadsheets/d/11jlo9UeWxZGwunhDb24hZBwAKc5b8ZKM9AYNWZaUyZY/edit#gid=243942212}
}
\description{
All Subsites as currently named in Qualtrics Reference Survey
}
