% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_monkey.R
\docType{data}
\name{survey_monkey}
\alias{survey_monkey}
\title{A survey monkey text selection}
\format{
A data frame with 87 rows and 1 variable:
\describe{
 \item{`What is the learning from this course that you are most excited about trying out?`}{A deidentified review in each row}
 }
 @source \url{https://www.surveymonkey.com/r/TLendofcourse}
}
\usage{
survey_monkey
}
\description{
Survey monkey reviews from the course survey in response to the question 
What is the learning from this course that you are most excited about trying out?, as a data frame.
}
\keyword{datasets}
