% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{tl_pal_blue}
\alias{tl_pal_blue}
\title{TL Default Blue Palette}
\format{
An object of class \code{character} of length 8.
}
\usage{
tl_pal_blue
}
\description{
TL Default Blue Palette
}
\keyword{datasets}
