% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{tl_palette2}
\alias{tl_palette2}
\title{Teaching Lab Color Palette Maker}
\usage{
tl_palette2(n = 6, base_color_start = NULL, end_color_start = NULL)
}
\arguments{
\item{n}{number of colors to generate}

\item{base_color_start}{the base color of the palette to start with, ramping from the color option}

\item{base_color_end}{the base color of the palette to end with, ramping towards the color option}
}
\value{
color ramp palette function
}
\description{
Teaching Lab Color Palette Maker
}
