% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\name{tl_select_percent}
\alias{tl_select_percent}
\title{Get the percent of a column that equals specific values}
\usage{
tl_select_percent(data, percent_equal)
}
\arguments{
\item{data}{the data for the plotter to use, should include all columns of interest}

\item{percent_equal}{string inputs to find the percent of the column that equals those values}
}
\value{
a percentage as a string
}
\description{
Automatically scaled stacked bar chart with TL theming
}
