% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_makers.R
\name{highlight_fun}
\alias{highlight_fun}
\title{Word highlighting}
\usage{
highlight_fun(data, highlight = TeachingLab::find_highlight(data))
}
\arguments{
\item{data}{the data to highlight}

\item{highlight}{the words to highlight}
}
\value{
a vector of strings
}
\description{
Provides html formatted highlighting
}
