% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\docType{data}
\name{no_data_plot_currently}
\alias{no_data_plot_currently}
\title{No Data Plot}
\format{
An object of class \code{ggplot2::ggplot} (inherits from \code{ggplot}, \code{ggplot2::gg}, \code{S7_object}, \code{gg}) of length 1.
}
\usage{
no_data_plot_currently
}
\description{
A plot that says no data available yet this year
}
\keyword{datasets}
