% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_functions.R
\docType{data}
\name{no_data_plot_filters}
\alias{no_data_plot_filters}
\title{No Data Plot}
\format{
An object of class \code{gg} (inherits from \code{ggplot}) of length 11.
}
\usage{
no_data_plot_filters
}
\description{
A plot that says no data available, and to check your filters
}
\keyword{datasets}
